moved vdifLinear2circular to !OLD_IPP area
The  for(int i=0...  is c++, not legal to gcc with the current
CPPFLAGS and -std=c99 fixed that problem but (for some odd reason)
did not then notice M_PI.  Leaving this puzzle for now for the folks
who actually need this.  (vdifio is critical, so not having it build
is a problem).


git-svn-id: https://svn.atnf.csiro.au/difx@10118 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876